🚚 Rename: Step컴포넌트 파일명 변경
